move reader functions to its own file
